cached_data removed, data_store refactored, get_value in data_store removed, prepare functions moved to utils
