+++ tests/scripts/test_data_store.py
@@ -57,8 +57,8 @@ def test_data_store_1():
     key = "input"
     value = backend.array([[1.0, 2, 3]])
     pm.data_store.add_static_data(key, value)
-    assert pm.data_store._cached_data.keys() == {"input"}
-    assert (pm.data_store._cached_data[key].value == value).all()  # type: ignore [union-attr]
+    assert pm.data_store.data_values.keys() == {"input"}
+    assert (pm.data_store.data_values[key].value == value).all()  # type: ignore [union-attr]
     assert pm.data_store._runtime_static_keys == set()
     assert pm.data_store._intermediate_non_differentiables._table == dict()
     assert pm.data_store.unused_keys == set()
@@ -86,12 +86,12 @@ def test_data_store_1_numpy():
     key = "input"
     value = backend.array([[1.0, 2, 3]])
     pm.data_store.add_static_data(key, value)
-    assert pm.data_store._cached_data.keys() == {
+    assert pm.data_store.data_values.keys() == {
         "input",
         "_MatrixMultiply_0_output_cache",
         "output_cache",
     }
-    assert (pm.data_store._cached_data[key].value == value).all()  # type: ignore[union-attr]
+    assert (pm.data_store.data_values[key].value == value).all()  # type: ignore[union-attr]
     assert pm.data_store._runtime_static_keys == set()
     assert pm.data_store._intermediate_non_differentiables._table == dict()
     assert pm.data_store.unused_keys == set()
@@ -106,9 +106,9 @@ def test_data_store_3():
         "w": backend.array([[1.0, 1, 1]]),
     }
     pm = mithril.compile(model, backend=backend, constant_keys=static_data)
-    assert pm.data_store._cached_data.keys() == {"_MatrixMultiply_1_output"}
+    assert pm.data_store.data_values.keys() == {"_MatrixMultiply_1_output"}
     assert (
-        pm.data_store.get_value("_MatrixMultiply_1_output") == backend.array(6.0)
+        pm.data_store.data_values["_MatrixMultiply_1_output"] == backend.array(6.0)
     ).all()  # type: ignore[union-attr]
     assert pm.data_store._runtime_static_keys == set()
     assert pm.data_store._intermediate_non_differentiables._table == dict()
@@ -200,7 +200,7 @@ def test_data_store_7():
     pm = mithril.compile(model, backend=backend, constant_keys={"input": value})
     res = pm.evaluate()
 
-    assert pm.data_store._cached_data.keys() == {"input"}
+    assert pm.data_store.data_values.keys() == {"input"}
     assert (res["output"] == value).all()  # type: ignore[union-attr]
     assert pm.data_store._runtime_static_keys == set()
     assert pm.data_store._intermediate_non_differentiables._table == dict()
@@ -216,8 +216,8 @@ def test_data_store_8():
     value = backend.array([[1.0, 2, 3]])
     pm = mithril.compile(model, backend=backend, constant_keys={"input": value})
 
-    assert pm.data_store._cached_data.keys() == {"output1"}
-    assert (pm.data_store.get_value("output1") == backend.sigmoid(value)).all()  # type: ignore[union-attr]
+    assert pm.data_store.data_values.keys() == {"output1"}
+    assert (pm.data_store.data_values["output1"] == backend.sigmoid(value)).all()  # type: ignore[union-attr]
     assert pm.data_store._runtime_static_keys == set()
     assert pm.data_store._intermediate_non_differentiables._table == dict()
     assert pm.data_store.unused_keys == {"input"}
@@ -233,8 +233,8 @@ def test_data_store_9():
     value = backend.array([[1.0, 2, 3]])
     pm = mithril.compile(model, backend=backend, constant_keys={"input": value})
 
-    assert pm.data_store._cached_data.keys() == {"output1"}
-    assert (pm.data_store.get_value("output1") == backend.sigmoid(value)).all()  # type: ignore[union-attr]
+    assert pm.data_store.data_values.keys() == {"output1"}
+    assert (pm.data_store.data_values["output1"] == backend.sigmoid(value)).all()  # type: ignore[union-attr]
     assert pm.data_store._runtime_static_keys == set()
     assert pm.data_store._intermediate_non_differentiables._table == dict()
     assert pm.data_store.unused_keys == {"input"}
@@ -250,8 +250,8 @@ def test_data_store_10():
     value = backend.array([[1.0, 2, 3]])
     pm = mithril.compile(model, backend=backend, constant_keys={"input": value})
 
-    assert pm.data_store._cached_data.keys() == {"input", "output2"}
-    assert (pm.data_store.get_value("output2") == backend.sigmoid(value)).all()  # type: ignore[union-attr]
+    assert pm.data_store.data_values.keys() == {"input", "output2"}
+    assert (pm.data_store.data_values["output2"] == backend.sigmoid(value)).all()  # type: ignore[union-attr]
     assert pm.data_store._runtime_static_keys == set()
     assert pm.data_store._intermediate_non_differentiables._table == dict()
     assert pm.data_store.unused_keys == set()
@@ -266,9 +266,9 @@ def test_data_store_11():
     value = backend.array([[1.0, 2, 3]])
     pm = mithril.compile(model, backend=backend, constant_keys={"input": value})
 
-    assert pm.data_store._cached_data.keys() == {"output1", "output3"}
-    assert (pm.data_store.get_value("output1") == backend.sigmoid(value)).all()  # type: ignore[union-attr]
-    assert (pm.data_store.get_value("output3") == backend.sigmoid(value) + 2).all()  # type: ignore[union-attr]
+    assert pm.data_store.data_values.keys() == {"output1", "output3"}
+    assert (pm.data_store.data_values["output1"] == backend.sigmoid(value)).all()  # type: ignore[union-attr]
+    assert (pm.data_store.data_values["output3"] == backend.sigmoid(value) + 2).all()  # type: ignore[union-attr]
     assert pm.data_store._runtime_static_keys == set()
     assert pm.data_store._intermediate_non_differentiables._table == dict()
     assert pm.data_store.unused_keys == {
@@ -294,12 +294,12 @@ def test_data_store_13():
         model, backend=backend, constant_keys={"left": left, "right": right}
     )
 
-    assert pm.data_store._cached_data.keys() == {"out"}
+    assert pm.data_store.data_values.keys() == {"out"}
     assert pm.data_store._runtime_static_keys == set()
     assert pm.data_store._intermediate_non_differentiables._table == dict()
     assert pm.data_store.unused_keys == {"left", "right"}
 
-    infered_value = pm.data_store.get_value("out")
+    infered_value = pm.data_store.data_values["out"]
     assert isinstance(infered_value, backend.DataType)
     np.testing.assert_allclose(infered_value, left + right, 1e-6)
 
@@ -328,7 +328,7 @@ def test_data_store_14():
         backend=backend,
         constant_keys={"input1": input1, "input2": input2, "kernel": kernel},
     )
-    assert pm.data_store._cached_data.keys() == {"input1", "out2"}
+    assert pm.data_store.data_values.keys() == {"input1", "out2"}
     assert pm.data_store._runtime_static_keys == set()
     assert pm.data_store._intermediate_non_differentiables._table == dict()
 
@@ -352,7 +352,7 @@ def test_data_store_14():
         "_Convolution2D_4_padding",
     }
 
-    infered_value = pm.data_store.get_value("out2")
+    infered_value = pm.data_store.data_values["out2"]
 
     assert isinstance(infered_value, backend.DataType)
     np.testing.assert_allclose(infered_value, backend.ones(1, 10, 15, 15) * 72, 1e-6)
@@ -382,7 +382,7 @@ def test_data_store_15():
         backend=backend,
         constant_keys={"input1": input1, "input2": input2, "kernel": kernel},
     )
-    assert pm.data_store._cached_data.keys() == {"input1", "out2"}
+    assert pm.data_store.data_values.keys() == {"input1", "out2"}
     assert pm.data_store._runtime_static_keys == set()
     assert pm.data_store._intermediate_non_differentiables._table == dict()
 
@@ -406,7 +406,7 @@ def test_data_store_15():
         "_Convolution2D_4_start",
     }
 
-    infered_value = pm.data_store.get_value("out2")
+    infered_value = pm.data_store.data_values["out2"]
 
     assert isinstance(infered_value, backend.DataType)
     np.testing.assert_allclose(infered_value, backend.ones(1, 10, 15, 15) * 72, 1e-6)
@@ -430,7 +430,7 @@ def test_data_store_16():
         safe_names=True,
     )
 
-    assert pm.data_store._cached_data.keys() == {
+    assert pm.data_store.data_values.keys() == {
         "axes",
         "_Transpose_0_output_cache",
         "_MatrixMultiply_1_output_cache",
@@ -462,7 +462,7 @@ def test_data_store_17():
         safe_names=False,
     )
 
-    assert pm.data_store._cached_data.keys() == {"_Add_0_output_cache", "output_cache"}
+    assert pm.data_store.data_values.keys() == {"_Add_0_output_cache", "output_cache"}
     assert pm.data_store._runtime_static_keys == {"right"}
     assert pm.data_store._intermediate_non_differentiables._table.keys() == set()
     assert pm.data_store.unused_keys == set()
@@ -492,7 +492,7 @@ def test_data_store_18():
         safe_names=True,
     )
 
-    assert pm.data_store._cached_data.keys() == set()
+    assert pm.data_store.data_values.keys() == set()
     assert pm.data_store._runtime_static_keys == set()
     assert pm.data_store._intermediate_non_differentiables._table.keys() == set()
     assert pm.data_store.unused_keys == set()
@@ -522,7 +522,7 @@ def test_data_store_19():
         safe_names=True,
     )
 
-    assert pm.data_store._cached_data.keys() == set()
+    assert pm.data_store.data_values.keys() == set()
     assert pm.data_store._runtime_static_keys == set()
     assert pm.data_store._intermediate_non_differentiables._table.keys() == set()
     assert pm.data_store.unused_keys == set()
@@ -556,7 +556,7 @@ def test_data_store_20():
     conn = model.conns.get_con_by_metadata(shp.output.metadata)
     shp_output_name = generated_keys[conn.key] if conn is not None else None
 
-    assert pm.data_store._cached_data.keys() == {"tensor_out"}
+    assert pm.data_store.data_values.keys() == {"tensor_out"}
     assert pm.data_store._runtime_static_keys == set()
     assert pm.data_store._intermediate_non_differentiables._table.keys() == set()
     assert pm.data_store.unused_keys == {"left", shp_output_name}
